pyproject.toml
[project]
name = "openaiwhisper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "ffmpeg>=1.4",
    "openai>=1.102.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "yt-dlp>=2025.8.22",
]
